requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0
scikit-learn>=1.3.0
opencv-python>=4.8.0.74
matplotlib>=3.7.0
tqdm>=4.66.0
pillow>=10.0.0
scipy>=1.11.0
albumentations>=1.3.0
huggingface_hub>=0.17.0
transformers>=4.33.0
